+++ csgotm/__init__.py
@@ -1,6 +1,5 @@
-from .account import (Account)
 from .market import (Market)
 
 __author__ = 'wendland0d'
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 __email__ = 'wendland0d@gmail.com'