+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # v0.1.0
 - Account API
 - Setting up initial requirements
-- First commit
+- "First commit"
 - Uploaded on PyPi
+
+#v0.2.0
+
+- Refactored the structure of module
+- Added exceptions
+- API:
+    1. Added trades methods 
+    2. Changed some methods with new exceptions